+++ sdk/core/azure-core/tests/async_tests/test_rest_http_response_async.py
@@ -8,8 +8,10 @@
 # Thank you httpx for your wonderful tests!
 import io
 import pytest
-from azure.core.rest import HttpRequest
+from azure.core.rest import HttpRequest, AsyncHttpResponse
+from azure.core.rest._aiohttp import RestAioHttpTransportResponse
 from azure.core.exceptions import HttpResponseError
+from utils import readonly_checks
 
 @pytest.fixture
 def send_request(client):
@@ -295,3 +297,16 @@ async def test_text_and_encoding(send_request):
 #         files=files,
 #     )
 #     await send_request(request)
+
+def test_initialize_response_abc():
+    with pytest.raises(TypeError) as ex:
+        AsyncHttpResponse()
+    assert "Can't instantiate abstract class" in str(ex)
+
+@pytest.mark.asyncio
+async def test_readonly(send_request):
+    """Make sure everything that is readonly is readonly"""
+    response = await send_request(HttpRequest("GET", "/health"))
+
+    assert isinstance(response, RestAioHttpTransportResponse)
+    readonly_checks(response)