+++ sdk/core/azure-core/CHANGELOG.md
@@ -10,8 +10,8 @@
 
 - The `text` property on `azure.core.rest.HttpResponse` and `azure.core.rest.AsyncHttpResponse` has changed to a method, which also takes
 an `encoding` parameter.
-- `azure.core.rest.HttpResponse` and `azure.core.rest.AsyncHttpResponse` are now protocols. They should not be initialized directly, instead
-your transport responses should inherit from them and implement their protocol.
+- `azure.core.rest.HttpResponse` and `azure.core.rest.AsyncHttpResponse` are now abstract base classes. They should not be initialized directly, instead
+your transport responses should inherit from them and implement them.
 
 ### Bugs Fixed